[macOS] Update the appearance of editable date/time controls to match the system https://bugs.webkit.org/show_bug.cgi?id=216616 <rdar://problem/69004603>
Reviewed by Devin Rousso.

Source/WebCore:

The default appearance of editable components within the date/time
control should closely match NSDatePicker.

1. Focused components should have rounded corners

This is achieved by adding the border-radius property to the
datetime-edit fields.

2. Use a raised colon separator for time inputs

The colon should be raised as it is surrounded two numbers. However,
since the colon is rendered in its own <div>, it remains lowered. To
raise the colon, the "ss03" font-feature-setting is enabled. Note that
this can cause issues with fonts that use "ss03" for a different purpose.
However, since we do not have support for the @font-feature-values rule
in CSS, there is currently no way to workaround this problem. Instead,
the issue should be fixed once we gain CSS support.

3. Components should have a fixed width, and the width of the control should match the content

Since the control is not guaranteed to use a monospace font, the current
design results in editable components being resized each time the value
changes. To fix this issue, the display type was changed to inline-block
and the min-width of a component is computed at runtime. See below for
more details.

* css/html.css:

Added UA stylesheet code that was removed from RenderTheme.

(input::-webkit-datetime-edit):
(input::-webkit-datetime-edit-year-field,):

Display inline-block allows us to set a fixed width for individual
components. The text is center-aligned so that changes are not
jarring. Finally, the font property is now "inherit !important",
as it does not make sense for the font used in a component to
differ from the control's font.

(input::-webkit-datetime-edit-year-field:focus,): Add rounded corners on focus.
* html/shadow/DateTimeFieldElement.cpp:
(WebCore::DateTimeFieldElement::DateTimeFieldElement):
* html/shadow/DateTimeFieldElements.cpp: Call setHasCustomStyleResolveCallbacks() to compute min-width.
(WebCore::DateTimeMillisecondFieldElement::DateTimeMillisecondFieldElement): Drive-by fix - milliseconds component has 3 digits.
* html/shadow/DateTimeNumericFieldElement.cpp:
(WebCore::DateTimeNumericFieldElement::resolveCustomStyle):

The min-width of a numeric field should be the largest possible width
depending on the font, the largest digit character in the font and the
length of the field. Ten possible values are tested (0-9).

* html/shadow/DateTimeNumericFieldElement.h:
* html/shadow/DateTimeSymbolicFieldElement.cpp:
(WebCore::DateTimeSymbolicFieldElement::resolveCustomStyle):

The min-width of a symbolic field should be the largest symbol given
a font. Two (meridiem) to twelve (month) values may be tested.

* html/shadow/DateTimeSymbolicFieldElement.h:
* rendering/RenderTheme.cpp:
* rendering/RenderTheme.h: Moved code to html.css.
* rendering/RenderThemeMac.h:
* rendering/RenderThemeMac.mm: Moved code to html.css.
* style/UserAgentStyle.cpp:
(WebCore::Style::UserAgentStyle::ensureDefaultStyleSheetsForElement): Moved code to html.css.
* style/UserAgentStyle.h:

LayoutTests:

Updated stylesheets so that font properties are not set on inner elements
of the control and rebaselined existing tests.

* fast/forms/date/date-editable-components/date-editable-components-focus-and-blur-events.html:
* fast/forms/date/date-editable-components/date-editable-components-mouse-events.html:
* fast/forms/date/date-pseudo-elements.html:
* fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-focus-and-blur-events.html:
* fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-mouse-events.html:
* fast/forms/time/time-editable-components/time-editable-components-focus-and-blur-events.html:
* fast/forms/time/time-editable-components/time-editable-components-mouse-events.html:
* platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/widgets/appearance/default-styles-expected.txt:
* platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/widgets/baseline-alignment-and-overflow.tentative-expected.txt:

The newly failing tests for <input type=week> are a false negative.
Previously, the style for this input type was empty, causing some
tests to incorrectly pass. The new failures are consistent with the
other date/time inputs, see the rest of the expectations for specifics.

* platform/mac-catalina/fast/forms/date/date-input-rendering-basic-expected.txt:
* platform/mac-catalina/fast/forms/time/time-input-rendering-basic-expected.txt: Added.
* platform/mac-mojave/fast/forms/date/date-input-rendering-basic-expected.txt:
* platform/mac-mojave/fast/forms/time/time-input-rendering-basic-expected.txt: Added.
* platform/mac-wk2/fast/forms/date/date-input-rendering-basic-expected.txt:
* platform/mac-wk2/fast/forms/date/date-pseudo-elements-expected.txt:
* platform/mac-wk2/fast/forms/time/time-input-rendering-basic-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267701 268f45cc-cd09-0410-ab3c-d52691b4dbfc