Adjust heuristic for checking whether view reaches visually non-empty state https://bugs.webkit.org/show_bug.cgi?id=217400
Reviewed by Geoffrey Garen.

Source/WebCore:

We decide that view reaches visually non-empty state and fire corresponding events when main document is parsed
and no more content (font or css resources) is to be loaded. An edge case is when the page has an empty document
and depends on async scripts to fill content (like wpt.fyi), then we may fire the events too early. To mitigate
this issue, now we don't set the state if content is completely empty and there are scripts to be loaded. Also,
because lazily loading elements need rendering update to decide if they can be loaded, we should not wait for
scripts where there are such elements.

API Test: WebKit.FirstVisuallyNonEmptyMilestoneWithDeferredScriptInEmptyDocument

* page/FrameView.cpp:
(WebCore::FrameView::checkAndDispatchDidReachVisuallyNonEmptyState):

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit/deferred-script-load-in-empty-document.html: Added.
* TestWebKitAPI/Tests/WebKitCocoa/FirstVisuallyNonEmptyMilestone.mm:
(-[RenderingProgressNavigationDelegate _webView:renderingProgressDidChange:]):
(TEST):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268192 268f45cc-cd09-0410-ab3c-d52691b4dbfc