Array.prototype.sort's sortBucketSort accesses an array in an invalid way that can lead to incorrect results with indexed properties on the prototype chain https://bugs.webkit.org/show_bug.cgi?id=217634 <rdar://problem/69489404>
Reviewed by Yusuke Suzuki.

JSTests:

* stress/sort-should-not-look-up-prototype-properties.js: Added.
(assert):

Source/JavaScriptCore:

Inside one of Array.prototype.sort's builtin helper methods, we are using an
array as an internal data structure to do some bookkeeping. However, we were
accessing it in such a way that it was reading properties from the prototype
chain. The code is written in a way such that it is only correct if it is
reading self properties (as the prototype chain can be user controlled). The
fix is to set this bookkeeping array's __proto__ to null, so we don't read
from the prototype chain.

* builtins/ArrayPrototype.js:
(globalPrivate.sortBucketSort):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268375 268f45cc-cd09-0410-ab3c-d52691b4dbfc