Lessen the reliance on VPATH in WebCore/DerivedSources.make https://bugs.webkit.org/show_bug.cgi?id=217696 <rdar://problem/70279941>
Reviewed by Darin Adler.

Due to the way vpath/VPATH works in `make`, we can get into a
situation where necessarily IDL files are not copied to their correct
locations.

WebKit comes with many IDL files. When building for internal purposes,
Apple can add to or modify this set of IDL files. When creating
Engineering builds, these additional IDL files are copied to
.../WebKitBuild/<configuration>/usr/local/includes. When performing
Production builds, these IDL files are included in the SDK. This
arrangement can be seen in DerivedSources.make:

    vpath %.idl $(BUILT_PRODUCTS_DIR)/usr/local/include $(SDKROOT)/usr/local/include

In order to get these IDL files into a location where they can be
uniformly found and handled, there is a step in DerivedSources.make
that copies them to the current directory, which is
WebKitBuild/<configuration>/DerivedSources/WebCore:

    $(ADDITIONAL_BINDING_IDLS) : % : WebKitAdditions/%
    cp $< .

The IDL files specified in ADDITIONAL_BINDING_IDLS are specified as
bare files names, meaning that vpath and VPATH are employed to find
them. Given the vpath specification previously shown, if the IDL files
can be found in $(BUILT_PRODUCTS_DIR)/usr/local/include, they are
copied locally. Otherwise, they should be found in
$(SDKROOT)/usr/local/include and copied locally.

As it turns out, vpath/VPATH resolution is performed not only on the
prerequisites of the build rule, but also the target (see section
4.5.3 of the GNU `make` manual). This means that the files specified
on the left side of the `cp` rule above are also searched for. And
since the standard IDL files (the ones that are being replaced with
the Apple-specific versions) can be found on VPATH as it's defined in
DerivedSources.make (specifically, in $(WebCore)/dom), then those
files participate in determining if the `cp` rule is executed.

Consider the normal build case: repositories are checked out (applying
the current modification timestamp to the files). During the build,
the files in $(ADDITIONAL_BINDING_IDLS) are copied to either
BUILT_PRODUCTS_DIR or to SDKROOT, which again modifies their
timestamps. We eventually get to the build rule for the `cp`
operation. Because the files in WebKitAdditions (be it in the local
build directory or the SDK) are *newer* than the checked-out ones in
$(WebCore)/dom due to their having been copied after they were
checked-out, then the `cp` command is invoked and then files are
copied to the current build directory.

However, consider a slightly different build case. In this case,
projects are checked out and built one at a time. So the project
producing the files in ADDITIONAL_BINDING_IDLS is checked out and then
the IDL files are copied to their intermediate location and have their
timestamps set to that time. Later, WebCore is checked out, and
DerivedSources.make is eventually invoked. Now, the files in
$(WebCore)/dom are *newer* than those in WebKitAdditions. `make`
determines that the IDL files are up-to-date and does not execute the
`cp` commands. The IDL files are not copied, and the build then either
fails (because of missing files) or proceeds incorrectly (because the
wrong IDL files are found).

The solution to this issue is to lessen the reliance on vpath/VPATH in
DerivedSources.make. Instead, be more explicit about where files can
be found and trim down vpath/VPATH to it's bare minimum. With
vpath/VPATH reduced, we remove the issue of files being accidentally
discovered that we don't want discovered.

The modifications effected to make this change are:

- Remove the definition of VPATH.
- Dynamically build up an idl-only "VPATH" as needed with the "vpath"
  directive.
- Manually look for the additional IDL files at their expected
  locations without using vpath/VPATH.
- For the files that used to be found on VPATH, specify their actual
  locations (typically by prefixing them with $(WebCore).
- (Cleanup) Consistently refer to local files with bare names (that
  is, remove the "./" that some files were prefixed with).
- (Cleanup) Use consistent spacing around "<" and ">" redirection
  symbols.
- (Cleanup) Remove some redundant prerequisites from the build rule
  involving IDL_INTERMEDIATE_PATTERNS.

No new tests -- no new or changed functionality.

* DerivedSources.make:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268566 268f45cc-cd09-0410-ab3c-d52691b4dbfc