TranslateTransformOperation shouldn't take in a FloatSize to convert z as a double https://bugs.webkit.org/show_bug.cgi?id=217246
Reviewed by Darin Adler.

The method used to get z as a double didn't use the FloatSize parameter, so we remove it. Since there already is a z()
method without an argument returning the underlying Length, we rename the method returning a double to make it clear
what type of value it's returning a double, and given how the call sites converts to a float eventually, we return
a float directly. We also follow the same pattern for x and y.

* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::getTransformFunctionValue):
* platform/graphics/transforms/TranslateTransformOperation.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268714 268f45cc-cd09-0410-ab3c-d52691b4dbfc